// FrontEnd-FinalProject/zipcode/src/models/employee.model.ts
export class Employee {
    id?: number = 0;
    firstname: string = '';
    lastname: string = '';
    rating: string = '';
    feedback: string = '';
    duration: string = '';
    improvements: string = '';
    completedItems:string = '';
    pendingItems:string= '';
    // jobExperience: number = 0;
    // salary: number = 0;
    // profile: string = '';
  }